Factor out tapdance.
