+++ users/talljoe/tapdance/tapdance.h
@@ -1,0 +1,8 @@
+#include "quantum.h"
+#include "td_setup.h"
+
+enum tap_dancers {
+  TD_SEMICOLON,
+  TD_GRAVE,
+};
+
