+++ users/talljoe/tapdance/td_setup.c
@@ -1,20 +1,4 @@
-//Tap Dance
-#include "talljoe.h"
-
-enum {
-  SINGLE_TAP = 1,
-  SINGLE_HOLD = 2,
-  DOUBLE_TAP = 3,
-  DOUBLE_HOLD = 4,
-  DOUBLE_SINGLE_TAP = 5, //send two single taps
-  TRIPLE_TAP = 6,
-  TRIPLE_HOLD = 7,
-  SPECIAL = 8
-};
-
-static struct {
-  int semicolon;
-} tap_state = {0};
+#include "tapdance.h"
 
 int cur_dance (qk_tap_dance_state_t *state) {
   if (state->count == 1) {
@@ -68,58 +52,3 @@ int hold_cur_dance (qk_tap_dance_state_t *state) {
   }
   else return SPECIAL;
 }
-
-// Send semi-colon + enter on two taps
-void tap_dance_semicolon_finished(qk_tap_dance_state_t *state, void *user_data) {
-  tap_state.semicolon = hold_cur_dance(state);
-  switch (tap_state.semicolon) {
-    case SINGLE_TAP: case DOUBLE_HOLD: register_code(KC_SCLN); break;
-    case SINGLE_HOLD: layer_on(_NUM); break;
-  }
-}
-
-void tap_dance_semicolon_reset(qk_tap_dance_state_t *state, void *user_data) {
-  switch (tap_state.semicolon) {
-    case SINGLE_TAP: case DOUBLE_HOLD: unregister_code(KC_SCLN); break;
-    case DOUBLE_TAP: {
-      if (get_mods()) {
-        SEND_STRING(";;"); // send normal when mods are pressed
-      }
-      else {
-        SEND_STRING(";\n");
-      }
-      break;
-    }
-    case TRIPLE_TAP: {
-      SEND_STRING(";\n\n");
-    }
-    case SPECIAL: layer_invert(_NUM); break;
-    case SINGLE_HOLD: layer_off(_NUM); break;
-  }
-  tap_state.semicolon = 0;
-}
-
-// Send `. ~. ```
-void tap_dance_grave_finished(qk_tap_dance_state_t *state, void *user_data) {
-  switch(state->count) {
-    case 1:
-      SEND_STRING("`");
-      break;
-    case 2:
-      SEND_STRING("~");
-      break;
-  }
-}
-
-void tap_dance_grave_each(qk_tap_dance_state_t *state, void *user_data) {
-  if(state->count == 3) {
-    SEND_STRING("```");
-  } else if (state->count > 3) {
-    SEND_STRING("`");
-  }
-}
-
-qk_tap_dance_action_t tap_dance_actions[] = {
-  [TD_SEMICOLON] = ACTION_TAP_DANCE_FN_ADVANCED(NULL, tap_dance_semicolon_finished, tap_dance_semicolon_reset),
-  [TD_GRAVE]     = ACTION_TAP_DANCE_FN_ADVANCED(tap_dance_grave_each, tap_dance_grave_finished, NULL),
-};
